Update debug-toolbars.md
I've found that ```return $provider``` does not return an array as expected by the ConfigAggregator. Instead, the static function  ```ConfigProvider::getConfig()``` returns the expected array.

Even then, that config array needs to be overwritten as shown above in order to avoid a 'Circular dependency detected while trying to resolve entry 'PhpMiddleware\PhpDebugBar\ConfigProvider'' error, which is caused by calls for ```$config = $container->get(ConfigProvider::class);``` within the debug bar package.
The only overwritten entry is '\PhpMiddleware\PhpDebugBar\ConfigProvider::class' which now returns the 'config' array from the container, which is what is required by ``$config = $container->get(ConfigProvider::class);```

In addition, aliases need to be added to the container required by the DebugBar as shown above, most likely within config/autoload/development.local.php. (They could also be added within config/autoload/dependencies.global.php if they are used by other services.) Then the PhpDebugBarMiddleware class is added to the pipeline only if `debug` is True in the application config.

Finally, I've included a simple example of how to use it within a request handler. A factory for the request handler in this example is not included but the debug bar still works. (Adding the debug bar to the factory is unnecessary in my opinion because all references to the debug bar are deleted in production..)

Thanks,
Collins.

Signed-off-by: Collins Pamba <collins@codeshop.co.ke>