Add note that PUT requests require the BodyParamsMiddleware
While building a small application that includes PUT routes, I came
across the need for including BodyParamsMiddleware in the route's
request pipeline, otherwise calls to getParsedBody return an empty
array. Given that I didn't find anything about this, specifically, in
the docs for the middleware, I created this small change to include it.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>