Update packages to maximum PHP 8.1 compatibility
This allows new dependency versions to be locked and in practice
makes all the packages that are in the `require` section compatible
with PHP 8.1.

Only 2 dev-require packages remain incompatible with this commit:
```sh
$ composer why-not php 8.1
mezzio/mezzio-aurarouter     3.1.0  requires  php (^7.3||~8.0.0)
mezzio/mezzio-laminasrouter  3.1.0  requires  php (^7.3 || ~8.0.0)
```

Prevents the tests for the lowest PHP 8.1 dependencies to fail by
indicating conflicting packages that are not directly required in
this library, but that they fail in some form under PHP 8.1.

For the Aura route package, the tests will exclude that adapter
for PHP versions 8.1 and newer.

Also updates the Psalm baseline.

Signed-off-by: Agustin Gomes <me@agustingomes.com>