Extend the delegator factor example
This change provides an an example file and method where the change
could happen, one that should help especially new developers to this
concept. For more experienced developers, they likely already know what
to do, and may do it in a different way. But for less-experienced
developers, or ones seeing the concept for the first time, this should
help avoid any confusion or uncertainty.