feat: adds support for Diactoros FilterServerRequestInterface
This patch adds support for the `FilterServerRequestInterface` in the `ServerRequestFactoryFactory`.
When the `FilterServerRequestInterface` service is defined, it will be used for the `$requestFilter` argument of Diactoros' `ServerRequestFactory::fromGlobals` factory.

Additionally, the patch adds a specialized factory for configuring a `FilterUsingXForwardedHeaders` to use for this service.
Users can configure it using the following configuration scheme:

```php
$config = [
    'laminas-diactoros' => [
        'server-request-filter' => [
            'x-forwarded-headers' => [
                'trusted-proxies' => string|string[],
                'trusted-headers' => string[],
            ],
        ],
    ],
    'dependencies' => [
        'factories' => [
            \Laminas\Diactoros\ServerRequestFilter\ServerRequestFilterInterface::class =>
                \Mezzio\Container\FilterUsingXForwardedHeadersFactory::class,
        ],
    ],
];
```

If they want to opt-out of usage of `X-Forwarded-*` headers entirely, they can now do so with the following configuration:

```php
$config = [
    'dependencies' => [
        'aliases' => [
            \Laminas\Diactoros\ServerRequestFilter\ServerRequestFilterInterface::class =>
                \Laminas\Diactoros\ServerRequestFilter\DoNotFilter::class,
        ],
    ],
];
```

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>