refactor: provide alternate container implementations during testing
Because we could get either a v1.1 or v2 version of psr/container at this point, we need to vary the type of InMemoryContainer used during testing based on the version detected.
This patch introduces a new variant, InMemoryContainerV2.
It also introduces a new interface, MutableMemoryContainerInterface, which both implementations implemenent.
The trait InMemoryContainerTrait is now composed in classes that previously called `new InMemoryContainer()`; they now use `$this->createContainer()` to create an instance, which will return one based on the psr/container version detected.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>