Update the Features section of the Getting Started guide
This change comprises a small set of changes for a series of different
reasons.

Firstly, as this is likely one of the first sections of the docs a
new/first-time reader may go to, the language has been reworked to be a
bit more sales-like and a little less technical. This is because this
section of the docs should really tell the reader what's in it for them,
encouraging them to keep exploring and learning.

Secondly, it changes the section's headers from bullet points to level
three headers. This makes them both easier to read and easier for the
reader to scan.

Thirdly, from the perspective of maintenance, sentences have been
changed to be all one one line, not split up over multiple lines and the
links have been extracted from being directly embedded with the text to
be reference-style links
(https://www.markdownguide.org/basic-syntax/#reference-style-links).
Personally, I find these easier to work with and maintain.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>